+++ src/pages/test/bottom-sheet/index.tsx
@@ -1,13 +1,27 @@
-import BottomSheet from '@/components/common/Modal/BottomSheet';
+import BottomSheet from '@/components/common/Modal/Bottom/BottomSheet';
 import useBottomSheet from '@/hooks/useBottomSheet';
 import useModal from '@/hooks/useModal';
 import useToast from '@/hooks/useToast';
+import { useEffect } from 'react';
 
 const BOTTOM_SHEET_ID = 'bottomSheetId';
 
 export default function BottomSheetPage() {
   const { onDragEnd, controls, isOpen, openHandler, closeHandler } = useBottomSheet();
   const { showToast, setPortalId } = useToast();
+
+  useEffect(() => {
+    console.log('useEffect');
+
+    const interval = setInterval(() => {
+      showToast({
+        status: 'success',
+        message: '토스트 입니다.',
+      });
+    }, 2000);
+
+    return () => clearInterval(interval);
+  }, [showToast]);
 
   return (
     <>
